@@ -42,21 +42,21 @@ pub trait SynchronousStreamCipher {
 }
 
 // TODO - Its a bit unclear to me why this is necessary
-impl SynchronousStreamCipher for Box<SynchronousStreamCipher + 'static> {
+impl SynchronousStreamCipher for Box<dyn SynchronousStreamCipher + 'static> {
     fn process(&mut self, input: &[u8], output: &mut [u8]) {
         let me = &mut **self;
         me.process(input, output);
     }
 }
 
-impl Encryptor for Box<SynchronousStreamCipher + 'static> {
+impl Encryptor for Box<dyn SynchronousStreamCipher + 'static> {
     fn encrypt(&mut self, input: &mut RefReadBuffer, output: &mut RefWriteBuffer, _: bool)
             -> Result<BufferResult, SymmetricCipherError> {
         symm_enc_or_dec(self, input, output)
     }
 }
 
-impl Decryptor for Box<SynchronousStreamCipher + 'static> {
+impl Decryptor for Box<dyn SynchronousStreamCipher + 'static> {
     fn decrypt(&mut self, input: &mut RefReadBuffer, output: &mut RefWriteBuffer, _: bool)
             -> Result<BufferResult, SymmetricCipherError> {
         symm_enc_or_dec(self, input, output)
